requirements.txt
boto3==1.5.7
botocore==1.8.21
certifi==2017.11.5
chardet==3.0.4
Django==1.11.5
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-filter==1.0.4
django-registration-redux==1.4
Django-Select2==5.11.1
django-storages==1.6.5
django-taggit==0.21.4
djangorestframework==3.6.4
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
nltk==3.2.5
pkg-resources==0.0.0
python-dateutil==2.6.1
pytz==2017.2
rake-nltk==1.0.1
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
textblob==0.13.0
urllib3==1.22
whitenoise==3.3.1
